+++ packages/core/src/fromEager.test.ts
@@ -67,7 +67,15 @@ test("types", () => {
 
   expectTypeOf(fromEager(() => "a" as const)).toEqualTypeOf<LazyPromise<"a">>();
 
+  expectTypeOf(fromEager(Promise.resolve("a" as const))).toEqualTypeOf<
+    LazyPromise<"a">
+  >();
+
   expectTypeOf(fromEager(() => box("a"))).toEqualTypeOf<LazyPromise<"a">>();
+
+  expectTypeOf(fromEager(Promise.resolve(box("a")))).toEqualTypeOf<
+    LazyPromise<"a">
+  >();
 
   expectTypeOf(fromEager(async () => "a" as const)).toEqualTypeOf<
     LazyPromise<"a">
@@ -119,6 +127,36 @@ test("types", () => {
   expectTypeOf(f("a" as const)).toEqualTypeOf<LazyPromise<"a">>();
 
   /* eslint-enable require-await */
+});
+
+test("no callback (resolve)", async () => {
+  const promise = fromEager(Promise.resolve("value"));
+  promise.subscribe(logSubscriber);
+  expect(readLog()).toMatchInlineSnapshot(`[]`);
+  await flushMicrotasks();
+  expect(readLog()).toMatchInlineSnapshot(`
+    [
+      [
+        "handleValue",
+        "value",
+      ],
+    ]
+  `);
+});
+
+test("no callback (reject)", async () => {
+  const promise = fromEager(Promise.reject("oops"));
+  promise.subscribe(logSubscriber);
+  expect(readLog()).toMatchInlineSnapshot(`[]`);
+  await flushMicrotasks();
+  expect(readLog()).toMatchInlineSnapshot(`
+    [
+      [
+        "handleError",
+        "oops",
+      ],
+    ]
+  `);
 });
 
 test("source is a plain value", () => {