+++ packages/core/src/fromEager.ts
@@ -33,13 +33,26 @@ class FromEagerSubscription implements InnerSubscription {
 
 class FromEagerProducer implements Producer<any> {
   constructor(
-    public callback: (options: { readonly signal: AbortSignal }) => any,
+    public arg:
+      | Promise<any>
+      | ((options: { readonly signal: AbortSignal }) => any),
   ) {}
 
   produce(innerSubscriber: InnerSubscriber<any>) {
+    if (this.arg instanceof Promise) {
+      this.arg.then(
+        (value) => {
+          innerSubscriber.resolve(value);
+        },
+        (error) => {
+          innerSubscriber.reject(error);
+        },
+      );
+      return;
+    }
     const innerSubscription = new FromEagerSubscription();
     // May throw.
-    const callbackReturn = (0, this.callback)(innerSubscription.options);
+    const callbackReturn = (0, this.arg)(innerSubscription.options);
     if (callbackReturn instanceof Promise) {
       callbackReturn.then(
         (value) => {
@@ -56,13 +69,19 @@ class FromEagerProducer implements Producer<any> {
 }
 
 /**
- * Converts a Promise to a LazyPromise. The callback can use an AbortSignal
- * passed in the options object.
+ * Converts a Promise to a LazyPromise. Takes either a Promise or a callback
+ * that may return one. A callback can use an AbortSignal passed in the options
+ * object.
  */
-export const fromEager = <Value>(
-  callback: (options: { readonly signal: AbortSignal }) => Value,
-): LazyPromise<
-  Value extends Promise<infer PromiseValue>
-    ? Flatten<PromiseValue>
-    : Flatten<Value>
-> => new LazyPromise<any>(new FromEagerProducer(callback));
+export const fromEager: {
+  <Value>(arg: Promise<Value>): LazyPromise<Flatten<Value>>;
+  <Value>(
+    arg: (options: { readonly signal: AbortSignal }) => Value,
+  ): LazyPromise<
+    Value extends Promise<infer PromiseValue>
+      ? Flatten<PromiseValue>
+      : Flatten<Value>
+  >;
+} = (
+  arg: Promise<any> | ((options: { readonly signal: AbortSignal }) => any),
+) => new LazyPromise<any>(new FromEagerProducer(arg));